requirements.txt
alembic==1.7.3
cachetools==4.2.4
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
google-auth==2.2.1
google-auth-oauthlib==0.4.6
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.5
MarkupSafe==2.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.19.0
pytube==11.0.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
SQLAlchemy==1.4.25
urllib3==1.26.7
Werkzeug==2.0.1
